String command in separate compilation unit
Allows application to determine linkage at link time rather than
preprocessor time. Or possibly not compile at all.
