+++ pickle.h
@@ -54,8 +54,10 @@ int pickle_eval(pickle_t *i, const char *t);
 int pickle_register_command(pickle_t *i, const char *name, pickle_command_func_t f, void *privdata);
 
 int pickle_set_result(pickle_t *i, const char *fmt, ...);
+int pickle_set_result_empty(pickle_t *i);
 int pickle_set_result_error(pickle_t *i, const char *fmt, ...);
 int pickle_set_result_error_arity(pickle_t *i, int expected, int argc, char **argv);
+int pickle_set_result_error_memory(pickle_t *i);
 int pickle_set_result_string(pickle_t *i, const char *s);
 int pickle_set_result_integer(pickle_t *i, long result);
 int pickle_get_result_string(pickle_t *i, const char **s);
@@ -66,7 +68,12 @@ int pickle_set_var_integer(pickle_t *i, const char *name, long r);
 int pickle_get_var_string(pickle_t *i, const char *name, const char **val);
 int pickle_get_var_integer(pickle_t *i, const char *name, long *val);
 
+unsigned long pickle_hash_string(const char *s, size_t len);
+int pickle_strcmp(const char *a, const char *b);
+
 int pickle_tests(void); /* returns: test passed || defined(NDEBUG) */
+
+int pickle_command_string(pickle_t *i, const int argc, char **argv, void *pd);
 
 #ifdef __cplusplus
 }
